Support React 19 ref as prop
This is a stab at supporting forwardRef and ref-as-prop (react 19).

By defaulting to CustomItemComponentProps ref type to div,
this *shouldn't* have breaking changes for most users.

Note: I also abstracted the types to encourage users to spread
onto the custom element. This may be helpful for changes passing
other attributes in the future?